datamodel: wsc_info's m1_len must be 16 bits
It will be used in a 16-bit value (1905.1 TLV length) and it can be much larger than 256.

+++ include/datamodel.h
@@ -309,7 +309,7 @@ struct radio {
      */
     struct {
         uint8_t   m1[1000]; /**< Buffer for constructing the M1 message. */
-        uint8_t   m1_len;   /**< Used length of @a m1. */
+        uint16_t  m1_len;   /**< Used length of @a m1. */
         uint8_t  *nonce;    /**< Pointer into @a m1 to location of the nonce. Length is 16 bytes. */
         uint8_t  *mac;      /**< Pointer into @a m1 to location of the MAC address. */
         uint8_t  *priv_key; /**< Private key, allocated separately. */
